Fixed related parts pricing - more info inside
Seems like having the related parts within the product details area made the price update occur on the same field when options changed. I removed out the related parts area from the product details section and the price ceased to update (good). I also moved the description tabs out to keep the order of elements on the page. Had to update styling slightly to keep this.